+++ applications/fishing-map/src/features/map/Map.tsx
@@ -23,8 +23,8 @@ import {
   Generators,
   LayerMetadataLegend,
 } from '@globalfishingwatch/layer-composer'
+import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import useMapInstance from 'features/map/map-context.hooks'
-import { UrlDataviewInstance } from 'types'
 import i18n from 'features/i18n/i18n'
 import { useClickedEventConnect, useMapTooltip, useGeneratorsConnect } from 'features/map/map.hooks'
 import { selectDataviewInstancesResolved } from 'features/workspace/workspace.selectors'