+++ applications/fishing-map/src/features/resources/resources.selectors.ts
@@ -1,49 +1,18 @@
 import { createSelector } from 'reselect'
-import {
-  Dataset,
-  DatasetTypes,
-  DataviewDatasetConfig,
-  EndpointId,
-} from '@globalfishingwatch/api-types'
-import { resolveEndpoint } from '@globalfishingwatch/dataviews-client'
+import { DatasetTypes, EndpointId, Resource } from '@globalfishingwatch/api-types'
+import { resolveDataviewDatasetResource } from '@globalfishingwatch/dataviews-client'
 import { Generators } from '@globalfishingwatch/layer-composer'
-import { ResourceQuery } from 'features/resources/resources.slice'
 import { selectDataviewInstancesResolved } from 'features/workspace/workspace.selectors'
-import { UrlDataviewInstance } from 'types'
 import { isGuestUser } from 'features/user/user.selectors'
 import { selectDebugOptions } from 'features/debug/debug.slice'
 import { ThinningLevels, THINNING_LEVELS } from 'data/config'
-
-export const resolveDataviewDatasetResource = (
-  dataview: UrlDataviewInstance,
-  { type, id }: { type?: DatasetTypes; id?: string }
-): {
-  dataset?: Dataset
-  datasetConfig?: DataviewDatasetConfig
-  url?: string
-} => {
-  if (!type && !id) return {}
-
-  const dataset = type
-    ? dataview.datasets?.find((dataset) => dataset.type === type)
-    : dataview.datasets?.find((dataset) => dataset.id === id)
-  if (!dataset) return {}
-  const datasetConfig = dataview?.datasetsConfig?.find(
-    (datasetConfig) => datasetConfig.datasetId === dataset.id
-  )
-  if (!datasetConfig) return {}
-  const url = resolveEndpoint(dataset, datasetConfig)
-  if (!url) return {}
-
-  return { dataset, datasetConfig, url }
-}
 
 export const selectDataviewsResourceQueries = createSelector(
   [selectDataviewInstancesResolved, isGuestUser, selectDebugOptions],
   (dataviewInstances, guestUser, { thinning }) => {
     if (!dataviewInstances) return
 
-    const resourceQueries: ResourceQuery[] = dataviewInstances.flatMap((dataview) => {
+    const resourceQueries: Resource[] = dataviewInstances.flatMap((dataview) => {
       if (dataview.config?.type !== Generators.Type.Track || dataview.deleted) {
         return []
       }
@@ -64,9 +33,7 @@ export const selectDataviewsResourceQueries = createSelector(
             return { ...datasetConfig, query: [...(datasetConfig.query || []), ...thinningQuery] }
           })
         }
-        const trackResource = resolveDataviewDatasetResource(dataview, {
-          type: DatasetTypes.Tracks,
-        })
+        const trackResource = resolveDataviewDatasetResource(dataview, DatasetTypes.Tracks)
         if (trackResource.url && trackResource.dataset && trackResource.datasetConfig) {
           trackQuery = {
             dataviewId: dataview.dataviewId as number,
@@ -77,11 +44,11 @@ export const selectDataviewsResourceQueries = createSelector(
         }
       }
 
-      const infoResource = resolveDataviewDatasetResource(dataview, { type: DatasetTypes.Vessels })
+      const infoResource = resolveDataviewDatasetResource(dataview, DatasetTypes.Vessels)
       if (!infoResource.url || !infoResource.dataset || !infoResource.datasetConfig) {
-        return trackQuery as ResourceQuery
+        return trackQuery as Resource
       }
-      const infoQuery: ResourceQuery = {
+      const infoQuery: Resource = {
         dataviewId: dataview.dataviewId as number,
         url: infoResource.url,
         datasetType: infoResource.dataset.type,