+++ packages/dataviews-client/src/index.ts
@@ -1,3 +1,13 @@
 export { default } from './dataviews-client'
-export { default as resolveDataviews } from './resolve-dataviews'
+export {
+  default as resolveDataviews,
+  resolveDataviewDatasetResource,
+  UrlDataviewInstance,
+  mergeWorkspaceUrlDataviewInstances,
+} from './resolve-dataviews'
+export {
+  getGeneratorConfig,
+  getDataviewsGeneratorConfigs,
+  MULTILAYER_SEPARATOR,
+} from './resolve-dataviews-generators'
 export { default as resolveEndpoint } from './resolve-endpoint'